third_party/blink/web_tests/external/wpt/scroll-animations/view-timelines/intermediate-transform-expected.txt

This is a testharness.js-based test.
[FAIL] View timeline delay
  assert_approx_equals: Unexpected value for startOffset expected 654 +/- 0.0001 but got 604
Harness: the test ran to completion.
